Collect common functionality into utility module.
