+++ handoff_response/python/irfsetup.py
@@ -19,36 +19,8 @@ import itertools
 import numpy as np
 import ROOT
 
-
-def get_branches(aliases):
-    """Get unique branch names from an alias dictionary."""
-    
-    ignore = ['pow', 'log10', 'sqrt', 'max']
-    branches = []
-    for k, v in aliases.items():
-
-        tokens = re.sub('[\(\)\+\*\/\,\=\<\>\&\!\-\|]', ' ', v).split()
-
-        for t in tokens:
-
-            if bool(re.search(r'^\d', t)) or len(t) <= 3:
-                continue
-
-            if bool(re.search(r'[a-zA-Z]', t)) and t not in ignore:
-                branches += [t]
-
-    return list(set(branches))
-
-
-def update_dict(d, u):
-    for k, v in u.items():
-        if isinstance(v, collections.Mapping):
-            r = update_dict(d.get(k, {}), v)
-            d[k] = r
-        else:
-            d[k] = u[k]
-    return d
-
+from irfutils import get_branches, get_cuts_from_xml, extract_generated_events
+from irfutils import get_class_types_from_xml, makedir
 
 def write_config(config, name):
 
@@ -59,48 +31,6 @@ def write_config(config, name):
         else:
             out_str += r"{}.{} = {}".format(name, k, v) + '\n'
     return out_str
-
-
-def extract_generated_events(config, meritdir):
-    """Extract the number of generated events in each MC sample.  If
-    'ngen' is not defined this method will try to extract this
-    information from the 'jobinfo' tree.
-
-    Parameters
-    ----------
-    config : dict
-        Configuration dictionary containing one dictionary for each MC
-        sample.
-    """
-
-    cfg_out = {}
-
-    for k, v in config.items():
-
-        if isinstance(v, dict):
-
-            cfg_out[k] = v.copy()
-
-            if 'ngen' in v:
-                continue
-            elif 'files' in v:
-                chain = ROOT.TChain('jobinfo')
-                for f in glob.glob(os.path.join(meritdir, v['files'])):
-                    chain.Add(f)
-                cfg_out[k]['ngen'] = getGeneratedEvents(chain)
-            else:
-                raise ValueError(r"Either 'ngen' or 'files' must be defined.")
-        elif isinstance(v, list):
-
-            ngen = eval(v[0])
-            logemin = eval(str(v[1]))
-            logemax = eval(str(v[2]))
-            cfg_out[k] = {'ngen': ngen, 'logemin': logemin, 'logemax': logemax}
-
-        else:
-            raise ValueError('Invalid type for MC dictionary element.')
-
-    return cfg_out
 
 
 def make_generated_events_str(config):
@@ -117,128 +47,19 @@ def make_generated_events_str(config):
     return out_str
 
 
-def getGeneratedEvents(chain):
-    NGen_sum = 0
-    vref = {}
+def makesetup(class_version, classname, variant, irfname, config, sel, livetime_bin=None):
 
-    vref['trigger'] = array.array('i', [0])
-    vref['generated'] = array.array('i', [0])
-    vref['version'] = array.array('f', [0.0])
-    vref['revision'] = array.array('f', [0.0])
-    vref['patch'] = array.array('f', [0.0])
-    chain.SetBranchAddress('trigger', vref['trigger'])
-    chain.SetBranchAddress('generated', vref['generated'])
-    chain.SetBranchAddress('version', vref['version'])
-
-    if chain.GetListOfBranches().Contains('revision'):
-        chain.SetBranchAddress('revision', vref['revision'])
-
-    if chain.GetListOfBranches().Contains('patch'):
-        chain.SetBranchAddress('patch', vref['patch'])
-
-    for i in range(chain.GetEntries()):
-        chain.GetEntry(i)
-        ver = int(vref['version'][0])
-        rev = int(vref['revision'][0])
-        patch = int(vref['patch'][0])
-        NGen = vref['generated'][0]
-        NGen_sum += NGen
-
-    return NGen_sum
-
-
-def strip(input_str):
-    return str(input_str.replace('\n', '').replace(' ', ''))
-
-
-def replace_aliases(cut_dict, aliases):
-
-    for k, v in cut_dict.items():
-        for k0, v0 in aliases.items():
-            cut_dict[k] = cut_dict[k].replace(k0, v0)
-
-
-def makedir(dirname):
-
-    if os.path.exists(dirname):
-        return
-
-    try:
-        os.mkdir(dirname)
-    except OSError as err:
-        if update:
-            pass  # do not care if direcotry exists
-        else:
-            raise err
-
-
-def get_class_types_from_xml(xmlfile):
-    xmldoc = minidom.parse(xmlfile)
-    class_version = xmldoc.getElementsByTagName(
-        'EventClass')[0].attributes['version'].value
-    itemlist = xmldoc.getElementsByTagName('EventMap')
-    classnames = [l.attributes['name'].value for l in itemlist[0].getElementsByTagName(
-        "EventCategory") if l.attributes['name'].value != "LLE"]
-    typenames = [l.attributes['name'].value for l in itemlist[1].getElementsByTagName(
-        "EventCategory")]
-    return class_version, classnames, typenames
-
-
-def get_cuts_from_xml(xmlfile):
-    """Extract event selection strings from the XML file."""
-
-    root = ElementTree.ElementTree(file=xmlfile).getroot()
-    event_maps = root.findall('EventMap')
-    alias_maps = root.findall('AliasDict')[0]
-
-    event_classes = {}
-    event_types = {}
-    event_aliases = {}
-
-    for m in event_maps:
-        if m.attrib['altName'] == 'EVENT_CLASS':
-            for c in m.findall('EventCategory'):
-                event_classes[c.attrib['name']] = strip(
-                    c.find('ShortCut').text)
-        elif m.attrib['altName'] == 'EVENT_TYPE':
-            for c in m.findall('EventCategory'):
-                event_types[c.attrib['name']] = strip(c.find('ShortCut').text)
-
-    for m in alias_maps.findall('Alias'):
-        event_aliases[m.attrib['name']] = strip(m.text)
-
-    replace_aliases(event_aliases, event_aliases.copy())
-    replace_aliases(event_aliases, event_aliases.copy())
-    replace_aliases(event_classes, event_aliases)
-    replace_aliases(event_types, event_aliases)
-
-    event_selections = {}
-    event_selections.update(event_classes)
-    event_selections.update(event_types)
-    event_selections.update(event_aliases)
-
-    return event_selections
-
-
-def makesetup(class_version, classname, variant, irfname, config, livetime_bin=None):
-
-    sel = config['selections']
-    generated = config['MC']
-
-    if not livetime_bin:
-        globfile = '*.root'
-        meritdir = config['meritdir']
+    if livetime_bin is None:
+        generated = config['MC']
     else:
-        meritdir = os.path.join(config['meritdir'], 'livetimebins')
-        globfile = '*%s*.root' % livetime_bin
+        generated = {livetime_bin : config['Livetime']['MC'][livetime_bin] }
 
     setup_string = """
 from IRFdefault import *
-import glob
 from math import *\n
 """
 
-    setup_string += "Prune.fileName = ''"    
+    setup_string += "Prune.fileName = ''\n"    
     #if livetime_bin is None:
     #    setup_string += "Prune.fileName = 'skim_%s.root'\n" % (irfname)
     #else:
@@ -253,19 +74,13 @@ from math import *\n
     branches += [config['Data']['var_xdir'], config['Data']['var_ydir'],
                  config['Data']['var_zdir'], config['Data']['var_energy']]
     setup_string += "Prune.branchNames = %s\n"%str(branches)
-    setup_string += """
-meritFiles = '%s'
-Data.files = sorted(glob.glob(meritFiles))\n
-""" % (os.path.join(meritdir, globfile))
 
-    if livetime_bin is None:
-        setup_string += make_generated_events_str(generated)
-    else:
-        n_generated = float(config['Livetime']['ngen'][livetime_bin])
-        setup_string += "Data.generated = [%g]\n" % n_generated
-        setup_string += "Data.logemin = [1.25]\n"
-        setup_string += "Data.logemax = [5.75]\n"
-
+    merit_files = []    
+    for k,v in generated.items():
+        merit_files += sorted(glob.glob(os.path.join(config['meritdir'],
+                                                     v['files'])))
+    setup_string += 'Data.files = %s\n'%str(merit_files)
+    setup_string += make_generated_events_str(generated)
     setup_string += """
 Data.friends = {}
 Data.tree_name = 'MeritTuple'\n
@@ -291,7 +106,8 @@ PSF.fit_pars={'ncore':[%g,1e-6,1.],
 """ % tuple(config['PSF']['fit_pars'][variant])
     setup_string += ('PSF.scaling_pars=%s\n' %
                      str(config['PSF']['scaling_pars'][variant]))
-
+    setup_string += ('Edisp.scaling_pars=%s\n' %
+                     str(config['Edisp']['scaling_pars'][variant]))    
     if livetime_bin is None:
         setup_string += "parameterFile = 'par_%s.root'\n\n" % (irfname)
         setup_string += "selectionName=\'%s\'\n\n" % irfname
@@ -299,8 +115,7 @@ PSF.fit_pars={'ncore':[%g,1e-6,1.],
         setup_string += "parameterFile = 'par_%s_%s.root'\n\n" % (
             irfname, livetime_bin)
         setup_string += "selectionName=\'%s_%s\'\n\n" % (irfname, livetime_bin)
-    setup_string += ('Edisp.scaling_pars=%s\n' %
-                     str(config['Edisp']['scaling_pars'][variant]))
+
 
     setup_string += r"""
 Bins.set_energy_bins()
@@ -341,14 +156,9 @@ def main():
 
     args = parser.parse_args()
 
-    default_bins = {'logemin': 0.75, 'logemax': 6.5, 'logedelta': 0.25,
-                    'cthmin': 0.2, 'cthmax': 1.0, 'cthdelta': 0.1,
-                    'edisp_energy_overlap': 0, 'edisp_angle_overlap': 0,
-                    'psf_energy_overlap': 0, 'psf_angle_overlap': 0}
-
     # Default configuration
     config = {}
-    config['selections'] = get_cuts_from_xml(args.classdefs)
+    selections = get_cuts_from_xml(args.classdefs)
     config.update(yaml.load(open(args.config, 'r')))
 
     if args.meritdir is not None:
@@ -361,32 +171,45 @@ def main():
 
     irf_version = config['irf_version']
     config['MC'] = extract_generated_events(config['MC'], config['meritdir'])
-
+    if 'Livetime' in config:
+        config['Livetime']['MC'] = extract_generated_events(config['Livetime']['MC'],
+                                                            config['meritdir'])
+        
+    
     for classname in classnames:
         for typename in typenames:
-            irfname = '%s_%s_%s_%s' % (class_version, classname,
-                                       irf_version, typename)
+            #irfname = '%s_%s_%s_%s' % (class_version, classname,
+            #                           irf_version, typename)
+            irfname = '%s_%s_%s' % (class_version, classname,
+                                    typename)
             print 'Doing', irfname
-            dirname = target_dir + '/' + irfname + '/'
+            dirname = os.path.join(target_dir, irfname)
             makedir(dirname)
-            f = open(dirname + 'setup.py', 'w')
-            f.write(makesetup(class_version, classname, typename, irfname, config))
-            f.close()
+            config['irf_name'] = str(irfname)
+            config['irf_class'] = str(classname)
+            config['irf_type'] = str(typename)
+            with open(os.path.join(dirname,'config.yaml'),'w') as f:
+                yaml.dump(config, f, Dumper=yaml.CDumper)
 
-    if args.livetime:
-        livetime_bins = config['Livetime']['ngen'].keys()
-
-        for classname in classnames:
-            for (typename, ltbin) in itertools.product(typenames, livetime_bins):
-                irfname = '%s_%s_%s_%s' % (
-                    class_version, classname, irf_version, typename)
-                print 'Doing', irfname, ltbin
-                dirname = target_dir + '/' + irfname + '/'
-                makedir(dirname)
-                f = open(dirname + 'setup_%s.py' % ltbin, 'w')
+            with open(os.path.join(dirname, 'setup.py'), 'w') as f: 
                 f.write(makesetup(class_version, classname,
-                                  typename, irfname, config, ltbin))
-                f.close()
+                                  typename, irfname, config, selections))
+
+        if not 'Livetime' in config:
+            continue
+                
+        livetime_bins = sorted(config['Livetime']['MC'].keys())
+
+        for (typename, ltbin) in itertools.product(typenames, livetime_bins):
+            #irfname = '%s_%s_%s_%s' % (
+            #    class_version, classname, irf_version, typename)
+            irfname = '%s_%s_%s' % (class_version, classname, typename)
+            print 'Doing', irfname, ltbin
+            dirname = os.path.join(target_dir, irfname)
+            f = open(os.path.join(dirname, 'setup_%s.py' % ltbin), 'w')
+            f.write(makesetup(class_version, classname,
+                              typename, irfname, config, selections, ltbin))
+            f.close()
 
 
 if __name__ == "__main__":
